pyproject.toml
[tool.poetry]
name = "django-local-read-write-replica"
version = "0.1.0"
description = "A simple django project with read/write db replica"
authors = ["Lucas Gigek <lucasgigek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
python-json-logger = "^2.0.1"
django-q = "^1.3.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-django = "^4.2.0"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''